Laser/Draw now work in simulator
I made a separate javascript file (LASER_CONTROL_mobile.js) that has
code to make the laser work using the iPhone simulator (not tested on a
real device yet). Currently it is necessary to edit index.html if you
want to switch back to using it in a browser, and change the script to
be the old LASER_CONTROL.js.