Conversion now creates proper index.html with all scripts.
hash_index.html serves as a template for all converted presentations.
It has references to all CSS and JS sources in /home/ec2-user/www/
which are shared by all hashed presentation directories.

The only thing that require customization is replacing NUM_SLIDES with
the actual number of slides in this presentation